Project Directory
Created anew FileTree class. Once a Folder is opened its path and all files that are in that project are displayed on the right hand side on the Project Directory window. Currently Displaying folder paths, needs to be simplified to just the file names. There may already be a function to do this or we may have to split the path by dilimeter and have the last string be displayed