General cleanup
Addedmore comments. Split the creationof each JPanel into its own functions and created seperate classes for the ActionListener. Made the JFrame and JPanels protected Global variables in order to access them in the ActionListeners. Also cleaned up the messages being displayed to the user whene they try to create a project with a name that already exists in the system.