Create New File
Lets the user creates a new file after a few checks are complete. Makes sure that a project folder is open for the file to be saved in. Makes sure that a file name already exists. Makes sure that the user actually gives a new file name (the .txt is not needed). Also did some more clean up of the Main class and changed some parameters around