File add on new project
Fixed the issue with creating a new file in a newly created project. Before you could not create a new file in a new project without first closing the text editor and reopening it. Now you are able to create new files in new projects without having to close the editor