Worked on User input
Project does not freeze up when program asks for user input. Still have to enter user input from the IDE Console. Need to find a way to redirect System.in to JTextField. Simple but unfortunetly does not seem to be working with System.in.available(). Currently only accepting user input from the IDE Console