Comments, Change in File Structure, addition of test projects for testing, and fixed create new issue
Added some comments to code. Got rid of redundant and unused code. Added some test projects for demenstration purposes. Fixed issue when creating a new project the files were not being saved correctly to the Project_Directory file, unless the project is closed and opened again. The current path was not being set correctly when creating a new project.